+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## `@svizzle/ui` v0.4.1 (next)
 
 - `Link`: fixed behaviour of `style` when `theme.color` is nullish
+- add `MultiBanner` component
 
 ## `@svizzle/atlas` v0.6.0 (next)